refactor: extract common logic to @utoo/pack-shared
- Move webpack compatibility logic to @utoo/pack-shared
- Move utility functions (processIssues, etc) to @utoo/pack-shared
- Update @utoo/pack to consume shared utilities
- Update @utoo/web to consume shared utilities
- Clean up unused files in @utoo/pack
