+++ packages/pack-shared/src/index.ts
@@ -1,3 +1,8 @@
 export { formatIssue, handleIssues, Issue } from "./issue";
 export { decodeMagicIdentifier } from "./magicIdentifier";
 export { renderStyledStringToErrorAnsi } from "./styledString";
+export * from "./config";
+export * from "./webpackCompat";
+export * from "./utils";
+
+
