+++ packages/pack/src/config/types.ts
@@ -1,0 +1,1 @@
+export * from "@utoo/pack-shared";
