+++ packages/pack/src/index.ts
@@ -1,15 +1,16 @@
-import { build } from "./build";
-import { serve } from "./dev";
-import * as webpackCompat from "./webpackCompat";
+import { build } from "./commands/build";
+import { serve } from "./commands/dev";
+import * as webpackCompat from "./config/webpackCompat";
 
 export { build };
 export { serve };
 
 const utoopack = { build, serve };
 export default utoopack;
-export * from "./find-root";
-export * from "./types";
-export * from "./webpackCompat";
+export * from "./config/types";
+export * from "./config/webpackCompat";
+export * from "./core/types";
+export * from "./utils/find-root";
 export type WebpackConfig = webpackCompat.WebpackConfig;
 namespace utoopack {
   export type WebpackConfig = webpackCompat.WebpackConfig;
