+++ packages/pack/src/utils/find-root.ts
