+++ packages/pack/src/utils/html-entry.ts
@@ -1,7 +1,7 @@
 import { DOMParser } from "domparser-rs";
 import fs from "fs";
 import path from "path";
-import { ConfigComplete, EntryOptions, HtmlConfig } from "../types";
+import { ConfigComplete, EntryOptions, HtmlConfig } from "../config/types";
 
 export function processHtmlEntry(config: ConfigComplete, projectPath: string) {
   if (!config.entry) return;
